+++ cphasing/prepare.py
@@ -96,8 +96,8 @@ def pipe(fasta, pairs, pattern="AAGCTT", min_mapq=0, min_contacts=3,
     if not skip_pairs2clm:
         cmd = ["cphasing-rs", "pairs2clm", str(pairs), "-c", str(min_contacts),
                 "-t", str(threads), "-o", f"{outprefix}.clm", "-q", str(min_mapq)]
-        if low_memory:
-            cmd.append("--low-memory")
+        # if low_memory:
+        #     cmd.append("--low-memory")
             
         flag = run_cmd(cmd, log=f'{log_dir}/prepare.pairs2clm.log')
         assert flag == 0, "Failed to execute command, please check log."